Added Checkpoint.get_default_score_fn to simplify best_model_handler creation
